Added Functionality to allow for lookup grid to only need to be created one time
the Lookup Grid only needs to be created the first time the program is run. This is a work around from the github file size limitations of just uploading the lookup grid directly to github. Now The program looks in the current directory for the lookup grid file and creates one if it does NOT find it....if it does find it...uses it and continues like normal